requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-taggit==1.2.0
humanize==0.5.1
idna==2.9
Pillow==7.1.2
python_dotenv
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.1.0
django-celery-beat==2.0.0
django-admin-honeypot
django-nested-inlines==0.1
django-phonenumber-field==5.0.0
webcolors==1.11.1
django-tinymce==3.1.0
sorl-thumbnail-serializer-field==0.2.1
djangorestframework-recursive==0.1.2
celery==4.4.4
django-colorfield==0.3.2